Migrated tests to respective modules
