+++ cli_test/helpers/fixtures.go
@@ -54,7 +54,7 @@ func NewFixtures(t *testing.T) *Fixtures {
 
 	buildDir := os.Getenv("BUILDDIR")
 	if buildDir == "" {
-		buildDir, err = filepath.Abs("../../build/")
+		buildDir, err = filepath.Abs("../../../../build/")
 		require.NoError(t, err)
 	}
